up to updating parameters

+++ src/main/java/hr/fer/zemris/bachelor/Constants/Constants.java
@@ -5,8 +5,8 @@ public class Constants {
     public static final int GREEDY_BREAKPOINT_BIT = 4;
     public static final int MAX_DATA_SIZE = 10000;
     public static final int MAXIMUM_TREE_DEPTH = 2;
-    public static final int ENVIRONMENT_REWARD_CORRECT = 1000;
-    public static final int ENVIRONMENT_REWARD_INCORRECT = 0;
+    public static final double ENVIRONMENT_REWARD_CORRECT = 1000;
+    public static final double ENVIRONMENT_REWARD_INCORRECT = 0;
     public static final double DONT_CARE_PROBABILITY = 0.33;
     public static final double MUTATION_PROBABILITY = 0.04;
