+++ src/main/java/hr/fer/zemris/bachelor/LCS/Environment/Environment.java
@@ -3,6 +3,6 @@ package hr.fer.zemris.bachelor.LCS.Environment;
 public interface Environment {
 
     boolean[] getInput();
-    int getReward(double action);
+    double getReward(double action);
 
 }
